extract optional tests to a separete file
