+++ tests/unit/test_data_model.py
@@ -1,17 +1,9 @@
 import copy
-from typing import Optional, Union
 
 import pytest
-from pydantic import BaseModel
 
-from datachain.lib.data_model import (
-    DataModel,
-    compute_model_fingerprint,
-    is_chain_type,
-    unwrap_optional,
-)
+from datachain.lib.data_model import DataModel, compute_model_fingerprint
 from datachain.lib.model_store import ModelStore
-from datachain.lib.utils import DataChainParamsError
 
 
 @pytest.fixture(autouse=True)
@@ -89,148 +81,3 @@ def test_compute_model_fingerprint_required_vs_optional_differs():
     fp_required = compute_model_fingerprint(Required, {"value": None})
     fp_optional = compute_model_fingerprint(OptionalField, {"value": None})
     assert fp_required != fp_optional
-
-
-# -- unwrap_optional -------------------------------------------------------
-
-
-class _Sample(BaseModel):
-    x: int
-
-
-@pytest.mark.parametrize(
-    ("annotation", "expected_inner", "expected_is_optional"),
-    [
-        (Optional[int], int, True),
-        (int | None, int, True),
-        (Union[int, None], int, True),
-        (Optional[str], str, True),
-        (Optional[_Sample], _Sample, True),
-        (Optional[list[int]], list[int], True),
-        (Optional[dict[str, int]], dict[str, int], True),
-        (int, int, False),
-        (str, str, False),
-        (list[int], list[int], False),
-        (_Sample, _Sample, False),
-        (Union[int, str], Union[int, str], False),
-    ],
-)
-def test_unwrap_optional(annotation, expected_inner, expected_is_optional):
-    inner, is_optional = unwrap_optional(annotation)
-    assert is_optional is expected_is_optional
-    assert inner == expected_inner
-
-
-def test_unwrap_optional_collapses_nested():
-    # Python collapses Optional[Optional[X]] to Optional[X], so this is mostly
-    # a sanity check that the helper still gives one-level unwrap.
-    inner, is_optional = unwrap_optional(Union[None, int])
-    assert is_optional is True
-    assert inner is int
-
-
-# -- is_chain_type with Optional ------------------------------------------
-
-
-def test_is_chain_type_optional_basic():
-    assert is_chain_type(Optional[int])
-    assert is_chain_type(int | None)
-    assert is_chain_type(Union[int, None])
-    assert is_chain_type(Optional[str])
-    assert is_chain_type(Optional[float])
-    assert is_chain_type(Optional[bool])
-
-
-def test_is_chain_type_optional_datamodel():
-    class M(DataModel):
-        x: int
-
-    assert is_chain_type(Optional[M])
-    assert is_chain_type(M | None)
-
-
-def test_is_chain_type_optional_collections():
-    assert is_chain_type(Optional[list[int]])
-    assert is_chain_type(Optional[dict[str, int]])
-    assert is_chain_type(list[int | None])  # nulls inside list (defer support)
-    assert is_chain_type(dict[str, int | None])
-
-
-# -- validate_default_none -----------------------------------------------
-
-
-def test_default_none_rejected_for_non_optional():
-    with pytest.raises(DataChainParamsError, match="default value `None` requires"):
-
-        class Bad(DataModel):
-            x: int = None  # type: ignore[assignment]
-
-
-def test_default_none_accepted_for_optional_typing():
-    class Ok(DataModel):
-        x: int | None = None
-
-    assert Ok().x is None
-
-
-def test_default_none_accepted_for_pep604():
-    class Ok(DataModel):
-        x: int | None = None
-
-    assert Ok().x is None
-
-
-def test_default_none_accepted_for_union_none():
-    class Ok(DataModel):
-        x: int | None = None
-
-    assert Ok().x is None
-
-
-def test_default_non_none_unaffected():
-    class Ok(DataModel):
-        x: int = 5
-        y: str = "hello"
-
-    assert Ok().x == 5
-    assert Ok().y == "hello"
-
-
-def test_default_none_error_names_field_and_type():
-    with pytest.raises(DataChainParamsError) as exc_info:
-
-        class Bad(DataModel):
-            count: int = None  # type: ignore[assignment]
-
-    assert "count" in str(exc_info.value)
-    assert "Bad" in str(exc_info.value)
-    assert "Optional[int]" in str(exc_info.value)
-
-
-# -- sentinel name does not collide with user fields ---------------------
-
-
-def test_is_null_field_name_allowed():
-    # `is_null` is a normal user field now; the sentinel uses the internal
-    # `_is_null` name, which pydantic forbids users from declaring.
-    class HasIsNull(DataModel):
-        is_null: bool = False
-        value: int = 0
-
-    assert HasIsNull(is_null=True).is_null is True
-
-
-def test_optional_datamodel_sentinel_does_not_collide_with_is_null_field():
-    from datachain.lib.signal_schema import SignalSchema
-
-    class M(DataModel):
-        is_null: bool = False  # same spelling as the old reserved word
-        value: int = 0
-
-    leaves = [
-        "__".join(p)
-        for p, _, has_subtree, _ in SignalSchema({"m": Optional[M]}).get_flat_tree()
-        if not has_subtree
-    ]
-    assert "m__is_null" in leaves  # the user's field
-    assert "m___is_null" in leaves  # the internal sentinel, no collision
